Allow anyone to use prf weapons
Previously, prf weapons in FE8 were still restricted to their class, so
I finally removed `prf` and zeroed out all the locks. I just reused code
from the FE7 item override.
