+++ randomizer/rom_editors/character_editor.py
@@ -45,13 +45,13 @@ class CharacterEditor:
                 weapon_type = self.get_weapon_type_for_class(new_class)
                 self._item_editor.load(character.item_pos, new_class, weapon_type)
                 self._item_editor.randomize()
-                self._item_editor.handle_overrides()
 
         self.randomize_palettes()
         self.add_promotions()
         self.fix_flyers()
 
         self.handle_overrides()
+        self._item_editor.handle_overrides()
 
         return self._rom_data
